be: created anew file for trigger related functions
